Move request_password_change to user_authn.
