--- openedx/core/djangoapps/user_api/forms.py
@@ -1,8 +1,0 @@
-# pylint: disable=unused-import
-# TODO: eventually move this implementation into the user_api
-"""
-Django Administration forms module
-"""
-from __future__ import absolute_import
-
-from student.forms import PasswordResetFormNoActive
